workspaces/lily_ai_review/backup/new_lily_ai_3/requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
python-dotenv>=1.0.0
supabase>=1.0.3
stripe>=5.4.0
pydantic>=2.0.0
jinja2>=3.1.2
python-multipart>=0.0.6
itsdangerous>=2.1.2
starlette>=0.27.0
email-validator>=2.0.0
